fix: 🔐 Split runtime auth into server-only auth.ts, CLI-only auth.cli.config.ts 🐛 Restore PRAGMA foreign_keys in drop-tables finally block 📝 Point auth:generate and docs at CLI config
