+++ package.json
@@ -62,7 +62,7 @@
     "e2e:codegen": "NODE_ENV=test playwright codegen http://localhost:3000",
     "typegen": "bun --bun run next typegen && tsc --noEmit",
     "env:typegen": "varlock typegen",
-    "auth:generate": "varlock run -- bun x auth@1.7.0-rc.1 generate --config ./src/lib/auth.config.ts --output ./src/db/schema/auth.ts --yes",
+    "auth:generate": "varlock run -- bun x auth@1.7.0-rc.1 generate --config ./src/lib/auth.cli.config.ts --output ./src/db/schema/auth.ts --yes",
     "db:generate": "varlock run -- bun x drizzle-kit generate",
     "db:migrate": "varlock run -- bun x drizzle-kit migrate",
     "db:push": "varlock run -- bun x drizzle-kit push --force",
